fixed unit tests
